allows 'touch NOERASE' option

+++ bin/start.csh
@@ -48,6 +48,8 @@ if (! -d "$datadir") then
 endif
 
 # Create list of subdirectories
+# If the file NOERASE exists, the old directories are not erased
+#   (start.x also knows then that var.dat is not created)
 foreach dir ($procdirs $subdirs)
   # Make sure a sufficient number of subdirectories exist
   set ddir = "$datadir/$dir"
@@ -58,19 +60,21 @@ foreach dir ($procdirs $subdirs)
     # when used with lnowrite=T, for example, we don't want to remove var.dat:
     set list = \
         `/bin/ls $ddir/VAR* $ddir/TAVG* $ddir/*.dat $ddir/*.info $ddir/slice*`
-    #if ($list != "") then
+    if (! -e NOERASE) then
       foreach rmfile ($list)
         if ($rmfile != $ddir/var.dat) rm -f $rmfile >& /dev/null
       end
-    #endif
+    endif
   endif
 end
 
 # Clean up previous runs
-if (-e $datadir/time_series.dat && ! -z $datadir/time_series.dat) \
-    mv $datadir/time_series.dat $datadir/time_series.`timestr`
-rm -f $datadir/*.dat $datadir/*.nml $datadir/param*.pro $datadir/index*.pro \
-      $datadir/averages/* >& /dev/null
+if (! -e NOERASE) then
+  if (-e $datadir/time_series.dat && ! -z $datadir/time_series.dat) \
+      mv $datadir/time_series.dat $datadir/time_series.`timestr`
+  rm -f $datadir/*.dat $datadir/*.nml $datadir/param*.pro $datadir/index*.pro \
+        $datadir/averages/* >& /dev/null
+endif
 
 # If local disk is used, copy executable to $SCRATCH_DIR of master node
 if ($local_binary) then
